build: relocate site netlify.toml to site/ directory
