+++ .github/workflows/verify-docs.yml
@@ -11,19 +11,22 @@ jobs:
     runs-on: ubuntu-latest
     timeout-minutes: 10
     steps:
-      - name: Git checkout
-        uses: actions/checkout@v4
-      - name: Use Node.js
-        uses: actions/setup-node@v4
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-node@v4
         with:
           node-version: lts/*
           cache: npm
-      - name: Install core dependencies
+
+      - name: Install dependencies
         run: npm ci --no-audit
-      - name: Install site dependencies
-        run: npm run site:build:install
+
+      - name: Install ./site dependencies
+        run: npm ci  --prefix=site --no-audit
+
       - name: Generate docs
-        run: npm run docs
+        run: npm run --prefix=site build
+
       - name: Check for changes
         run: |
           if [ -z "$(git status --porcelain)" ]; then
