+++ site/scripts/util/generate-command-data.js
@@ -1,5 +1,5 @@
-import { createMainCommand } from '../../../dist/commands/index.js'
-import { sortOptions } from '../../../dist/utils/command-helpers.js'
+import { createMainCommand } from '../../../src/commands/index.js'
+import { sortOptions } from '../../../src/utils/command-helpers.js'
 
 const program = createMainCommand()
